fix: dropping a consumer can block indefinitely
This is a stopgap fix. I've created the following issues as a follow up:

https://github.com/edenhill/librdkafka/issues/4041
https://github.com/fede1024/rust-rdkafka/issues/509

The call to `rd_kafka_consumer_close()` in the `Drop` trait of
`BaseConsumer` can block in the rebalance callback indefinitely. It's
stuck in the call `rd_kafka_rebalance_protocol()` waiting for a queue
operation to be completed. To avoid this rebalance operation we avoid
calling `rd_kafka_consumer_close()` explicitly and let it get called in
`rd_kafka_destroy_flags()`, with the flag
`RD_KAFKA_DESTROY_F_NO_CONSUMER_CLOSE` so that the rebalance callbacks
are not called. The other side effect is any outstanding offset state
will not be persisted to the broker.

Note: It's my understanding that this final rebalance operation is not
critical and the rebalance protocol will recover from a non-orderly
consumer shutdown.